Functions shifted to corresponding classes
