+++ ShiftString/src/App.java
@@ -33,7 +33,7 @@ public class App {
             boolean keepGoing = true;
             while (keepGoing) {
                 processSingleShift(scanner);
-                keepGoing = Console.askToContinue(scanner);
+                keepGoing = askToContinue(scanner);
             }
         }
 
@@ -46,10 +46,10 @@ public class App {
      */
     private static void processSingleShift(Scanner scanner) {
         // Input string
-        String inputString = Console.getValidatedStringInput(scanner);
+        String inputString = Validator.getValidatedStringInput(scanner);
 
         // Input shift
-        int shiftPositions = Console.getValidatedShiftInput(scanner);
+        int shiftPositions = Validator.getValidatedShiftInput(scanner);
 
         // Perform shift
         String result = Shifter.shiftString(inputString, shiftPositions);
@@ -58,5 +58,27 @@ public class App {
                 ", shift=" + shiftPositions + 
                 ", result length=" + result.length() + ")");
         Console.printMessage("\nShifted String: \"" + result + "\"");
+    }
+
+    /**
+     * Asks the user if they want to continue and validates the input.
+     * @param scanner the Scanner object for reading input
+     * @return true if the user wants to continue, false otherwise
+     */
+    static boolean askToContinue(Scanner scanner) {
+        while (true) {
+            Console.printMessage("\nDo you want to go again? (y/n): ");
+            String answer = scanner.nextLine().trim().toLowerCase();
+            switch (answer) {
+                case "y":
+                    return true;
+                case "n":
+                    return false;
+                default:
+                    Console.printMessage("Invalid input. Please enter 'y' for yes or 'n' for no.");
+                    Console.logWarning(answer + " is not a valid input for continuation prompt.");
+                    continue;
+            }
+        }
     }
 }
