+++ ShiftString/src/Validator.java
@@ -1,3 +1,4 @@
+import java.util.Scanner;
 import java.util.regex.Pattern;
 
 /**
@@ -10,10 +11,56 @@ public class Validator {
     private static final int MIN_SHIFT = -1000;
     private static final int MAX_SHIFT = 1000;
     private static final Pattern VALID_STRING_PATTERN = Pattern.compile("[a-zA-Z\\s]+");
-    public static final String SHIFT_RANGE_MESSAGE = "Shift positions must be between " + MIN_SHIFT + " and " + MAX_SHIFT;
+    private static final String SHIFT_RANGE_MESSAGE = "Shift positions must be between " + MIN_SHIFT + " and " + MAX_SHIFT;
     private static final String EMPTY_STRING_ERROR = "Input string cannot be null or empty";
     private static final String EXCEEDS_MAX_STRING_LENGTH_ERROR = "Input string must not exceed " + MAX_STRING_LENGTH + " characters";
     private static final String INVALID_STRING_PATTERN_ERROR = "Input string must contain only alphabetic letters (a-z, A-Z) and spaces";
+
+    /**
+     * Prompts the user for a valid integer input and validates it.
+     * @param scanner the Scanner object for reading input
+     * @return a validated integer
+     */
+    static int getValidatedShiftInput(Scanner scanner) {
+        Console.printMessage("\nEnter a number between -1000 and 1000 to shift your string by");
+        Console.printMessage("(negative = left shift, positive = right shift):");
+        while (true) {
+            String line = scanner.nextLine().trim();
+            if (line.isEmpty()) {
+                Console.logWarning("Empty input for shift positions.");
+                Console.printError("Please enter a valid integer for shift positions in range (-1000, 1000).");
+                continue;
+            }
+            try {
+                int shiftPositions = Integer.parseInt(line);
+                return Validator.validateInt(shiftPositions);
+            } catch (NumberFormatException e) {
+                Console.logWarning(line + " is not a valid integer input for shift positions.");
+                Console.printError("Please enter a valid integer for shift positions in range (-1000, 1000).");
+            } catch (IllegalArgumentException e) {
+                Console.logWarning("Illegal argument in getValidationShiftInput() " + e.getMessage());
+                Console.printError(e.getMessage());
+            }
+        }
+    }
+
+    /**
+     * Prompts the user for a valid string input and validates it.
+     * @param scanner the Scanner object for reading input
+     * @return a validated string
+     */
+    static String getValidatedStringInput(Scanner scanner) {
+        while (true) {
+            Console.printMessage("\nPlease enter a string with maximum 1000 characters including whitespaces: ");
+            var inputString = scanner.nextLine();
+            try {
+                return Validator.validateString(inputString);
+            } catch (IllegalArgumentException e) {
+                Console.logWarning("Illegal argument in getValidationStringInput() " + e.getMessage());
+                Console.printError(e.getMessage());
+            }
+        }
+    }
 
     /**
      * Validates the shift positions.
